backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pgvector==0.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.57.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
